Fix RF/BBB-ADC wiki content
If an image folder is uploaded containing a corrupted image file
it causes the corresponding image directory to not load correctly
into WikiJS, although it is correctly uploaded to the GitHub repo.
To correct this issue it is necessary to delete the whole directory
and re-upload teh whole folder and its images.